timer and signal repair
The timer that updates the weather data has been repaired, and a binding has been added to emit a signal when the weather forecast is updated so it is easier to manage the data.

In MenuRepresentation the timer to update the forecast data was eliminated, now it is managed from weatherDate.
issues #2